Bug 8367: Template fixes + fixes to C4/Reserves.pm and Koha/Hold.pm
Fixed waitingreserves.tt template so the borrowernumber is
successfully passed to waitingreserves.pl upon cancelling a hold, so
users can now successfully cancel holds from the waitingreserves.pl
interface.

Removed check for ReserveMaxPickupdateDelay in Koha/Hold.pm as this
syspref is removed in the previous patch. Instead the
expirationdate is calculated the same way as the lastpickupdatedate that is
retrieving the holdpickupwait value for the appropriate issuingrule and then
adding that onto the waiting date.

Same test plan as main patch

Sponsored-By: Brimbank Library, Australia